refactor URN and DOI Generator
